+++ README.md
@@ -1,12 +1,16 @@
 # Edible Carrot or Poisonous Hemlock?
 A crop-weed discrimination project for 16-824 Visual Learning &amp; Recognition
 
+Crop-weed discrimination is a critical challenge as weeds reduce yield while often being difficult to discriminate from crops by a non domain expert. This repository implements Pairwise Confusion for [Fine-Grained Visual Classification](https://www.ecva.net/papers/eccv_2018/papers_ECCV/papers/Abhimanyu_Dubey_Improving_Fine-Grained_Visual_ECCV_2018_paper.pdf) on the [Plant Seedlings Dataset](https://vision.eng.au.dk/plant-seedlings-dataset/#citation).
+
 ## Installation
 Install the required python packages:
 
 ```
 pip install -r requirements.txt
 ```
+
+Download the [Plant Seedlings Dataset](https://vision.eng.au.dk/?download=/data/WeedData/NonsegmentedV2.zip) and place in the `data` folder
 
 ## Training
 To train a new model, modify the hyperparameters in [config.yaml](config.yaml) and run the following command:
@@ -33,11 +37,11 @@ This will return the accuracy of the model and overwrite `confusion_matrix.png`,
 
 ## Pairwise Confusion Integration with ResNet
 
-This implementation integrates Pairwise Confusion with a modified ResNet model that outputs both features and logits during the forward pass. To train the model with Pairwise Confusion, use the `config_pc.yaml` configuration, which specifies the ResNet variant supporting feature extraction.
+This implementation integrates Pairwise Confusion with a modified ResNet model that outputs both features and logits during the forward pass.
 
 ### Loss Functions and Objectives
 
-The Pairwise Confusion loss is computed in `trainer.py` through two additional loss functions:
+The Pairwise Confusion loss is computed in [utils.py](crop_weed_discrim/utils/utils.py) through two additional loss functions:
 - `PairwiseConfusion(features)`: Encourages feature compactness within a batch, helping to prevent overfitting.
 - `EntropicConfusion(features)` : Counterbalances overconfidence by increasing the entropy of the model’s predictions, promoting a smoother and more robust output distribution.
 
@@ -48,4 +52,4 @@ These are combined with the standard classification loss (`CrossEntropyLoss`) to
 - `lambda_pc`: Controls the weight of the Pairwise Confusion loss.
 - `lambda_ec`: Controls the weight of the Entropic Confusion loss.
 
-Both hyperparameters can be adjusted in `trainer.py` to tune the influence of these losses on the training process.
+Both hyperparameters can be adjusted in [config.yaml](config.yaml) to tune the influence of these losses on the training process.
