--- config_pc.yaml
@@ -1,14 +1,0 @@
-model:
-  name: resnet_pc
-
-hyperparams:
-  epochs: 50
-  batch_size: 256
-  lr: 0.0001
-  step_size: 10
-  gamma: 0.35
-  use_cuda: True
-  val_every: 5
-
-dataset:
-  name: plant_seedlings
