+++ test.py
@@ -9,8 +9,7 @@ from torchvision import transforms
 
 from crop_weed_discrim import trainer, tester
 from crop_weed_discrim.utils.utils import ARGS
-# Import ResNetCNN that outputs both features and logits for PC
-from crop_weed_discrim.models.resnet_pc import ResNetCNN
+from crop_weed_discrim.models.resnet_cnn import ResNetCNN
 from crop_weed_discrim.utils.dataloader import PlantSeedlingsDataset, SubsetWrapper
 
 if __name__ == "__main__":
@@ -44,7 +43,10 @@ if __name__ == "__main__":
         step_size = config['hyperparams']['step_size'],
         gamma = config['hyperparams']['gamma'],
         use_cuda = config['hyperparams']['use_cuda'],
-        val_every = config['hyperparams']['val_every']
+        val_every = config['hyperparams']['val_every'],
+        pc_loss = config['pairwise_confusion']['pc_loss'],
+        lambda_pc = config['pairwise_confusion']['lambda_pc'],
+        lambda_ec = config['pairwise_confusion']['lambda_ec'],
     )
 
     print("Loaded config.yaml:")
